LayoutTests/sputnik/Conformance/15_Native_Objects/15.1_The_Global_Object/S15.1_A2_T1-expected.txt

S15.1_A2_T1

PASS TypeError: this is not a function. (In 'this()', 'this' is an instance of Window)

TEST COMPLETE
